requirements.txt
streamlit == 1.52.2
pandas == 2.3.3
bidict == 0.23.1
fpdf2 == 2.8.7
python-docx == 1.2.0